Add expect/actual for LinksManagerTest.
